api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Challenge for MaisTodos company"
authors = ["rafael rotiroti <rotirotirafa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
python-creditcard = {path = "../python-creditcard"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2 = "^2.9.9"
pydantic = "^2.5.1"
bcrypt = "^4.0.1"
cryptography = "^41.0.5"
python-jose = {extras = ["openssl"], version = "^3.3.0"}
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
